requirements.txt
alembic==0.9.3.dev0
appdirs==1.4.4
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.4.2
asyncpg==0.21.0
attrs==20.2.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
databases==0.3.2
decorator==4.1.2
defer==1.0.6
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
fastapi==0.61.1
FastAPI-SQLAlchemy==0.1.6
filelock==3.0.12
h11==0.9.0
httplib2==0.9.2
httptools==0.1.1
idna==2.6
importlib-metadata==1.7.0
iniconfig==1.0.1
ipython==5.5.0
ipython-genutils==0.2.0
isort==5.5.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
lxml==4.5.2
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==8.5.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
packaging==20.4
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
pipenv==2020.8.13
pluggy==0.13.1
postgres==3.0.0
prompt-toolkit==1.0.15
protobuf==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py==1.9.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.6
Pygments==2.2.0
pygobject==3.26.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pytest==6.0.2
pytest-asyncio==0.14.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.14.0
pytils==0.3
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
selectors==0.0.14
simplegeneric==0.8.1
simplejson==3.13.2
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
system-service==0.3
systemd-python==234
toml==0.10.1
traitlets==4.3.2
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
user-agent==0.1.9
uvicorn==0.11.8
uvloop==0.14.0
vboxapi==1.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
wadllib==1.3.2
wcwidth==0.1.7
weblib==0.1.30
websockets==8.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
